ta_agent/requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
streamlit
scipy
python-dotenv
pydantic-settings
certifi
requests
urllib3
ta
backtrader

# LangChain Core and dependencies (with compatible versions)
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-groq>=0.0.1
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
langgraph>=0.0.20

# AI APIs
openai>=1.0.0
anthropic
groq
google-generativeai>=0.3.0

# Vector stores
chromadb
tiktoken
faiss-cpu

# Database
sqlalchemy

# RAG Dependencies
sentence-transformers
pypdf
python-docx
unstructured[local-inference]
markdown
python-jose[cryptography]
passlib[bcrypt]
argon2-cffi
python-multipart

# Sentiment Analysis Dependencies
textblob==0.17.1
feedparser==6.0.10
requests[security]

# Database drivers (uncomment as needed)
psycopg2-binary  # PostgreSQL driver
asyncpg  # Async PostgreSQL driver
# pymysql  # MySQL driver
